+++ package.json
@@ -32,46 +32,35 @@
   },
   "homepage": "https://gitlab.com/Jhonsebas77/PokedexRN-API#README",
   "dependencies": {
-    "@types/supertest": "^2.0.6",
-    "body-parser": "^1.18.2",
+    "@types/bunyan": "^1.8.4",
+    "body-parser": "^1.18.3",
     "bunyan": "^1.8.12",
     "conf": "^1.4.0",
-    "config": "^1.30.0",
+    "config": "^2.0.1",
     "cors": "^2.8.4",
     "dotenv": "^5.0.1",
     "express": "^4.16.3",
     "express-healthcheck": "^0.1.0",
     "gulp": "^3.9.1",
-    "jest": "^23.6.0",
-    "method-override": "^2.3.10",
-    "mongoose": "^5.0.12",
-    "morgan": "^1.9.0",
-    "nconf": "^0.10.0",
-    "superagent": "^3.8.2",
-    "supertest": "^3.0.0",
-    "ts-node": "^5.0.1",
-    "typescript": "^2.8.3",
-    "@types/axios": "^0.14.0",
-    "@types/body-parser": "^1.17.0",
-    "@types/bunyan": "^1.8.4",
-    "@types/config": "0.0.34",
-    "@types/cors": "^2.8.4",
-    "@types/dotenv": "^4.0.3",
-    "@types/express": "^4.16.0",
-    "@types/gulp": "^4.0.5",
-    "@types/jest": "^22.2.2",
-    "@types/method-override": "0.0.31",
-    "@types/mongoose": "^5.2.18",
-    "@types/morgan": "^1.7.35",
-    "@types/nconf": "0.0.37",
-    "@types/node": "^10.11.4",
-    "@types/sugar": "^2.0.2",
-    "@types/superagent": "^3.5.7",
-    "nodemon": "^1.17.3",
-    "ts-jest": "^22.4.2"
+    "mongoose": "^5.2.14",
+    "morgan": "^1.9.0"
   },
   "devDependencies": {
-
+    "@types/body-parser": "^1.17.0",
+    "@types/config": "^0.0.34",
+    "@types/cors": "^2.8.4",
+    "@types/express": "^4.16.0",
+    "@types/jest": "^23.3.2",
+    "@types/mongoose": "^5.2.12",
+    "@types/node": "^10.9.4",
+    "@types/supertest": "^2.0.6",
+    "jest": "^23.5.0",
+    "nodemon": "^1.18.4",
+    "supertest": "^3.3.0",
+    "ts-jest": "^23.1.4",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.11.0",
+    "typescript": "^3.0.3"
   },
   "jest": {
     "transform": {
