--- src/Interfaces/IPokedex.ts
@@ -1,9 +1,0 @@
-export interface IPokedex {
-    name: String,
-    url: String,
-    urlSprite: String,
-    type: {
-        type1: String,
-        type2?: String
-    }
-}
