--- src/Interfaces/Iitem.ts
@@ -1,14 +1,0 @@
-export interface Iitem {
-    name: String,
-    urlSprite: String,
-    idDex: Number,
-    effect_entries: {},
-    held_by_pokemon: {},
-    category: String,
-    cost: Number,
-    fling: {},
-    machines: {},
-    baby_trigger_for: {},
-    evolution_for: {},
-    attributes: {}
-}
