--- src/Interfaces/IitemList.ts
@@ -1,7 +1,0 @@
-export interface IitemList {
-    name: String,
-    url: String,
-    urlSprite: String,
-    idDex: Number,
-    type: String
-}
