+++ src/api/routes/index.ts
@@ -1,0 +1,4 @@
+import { routesPokedex } from '../Pokedex/routes'
+export const allRoutes = [
+    routesPokedex
+]
