+++ src/core/app/index.ts
@@ -1,0 +1,7 @@
+import express from 'express'
+export class ExpressApp implements IData {
+    public middleware: express.Application
+    constructor() {
+        this.middleware = express()
+    }
+}
